+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Current
 
-- Feature: Added `eth_hentai.gas.node_default_gas_price_strategy` to support BNB Chain
 - Feature: Added revert reason extraction for failed transactions
+- Feature: Added `eth_hentai.gas.node_default_gas_price_strategy` to support BNB Chain
 - Fix: BNB Chain compatibility fixes because of brokeness in Ethereum JSON-RPC
 - Fix: Wait 10 seconds instead of 5 seconds to ganache-cli to launch, as the latter is too slow for some computers
+- Fix: Optimize `wait_transactions_to_complete`
 
 # 0.5